added crowdsale + CSS changes

+++ contracts/TokenAirDrop.sol
@@ -3,18 +3,17 @@ pragma solidity ^0.8.3;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
-
 // THIS IS NOT INTENDED FOR PRODUCTION USE, ONLY FOR TESTING
 contract TokenAirDrop is Ownable {
-    string[] claimed; 
-    uint claimedCount;
+    string[] claimed;
+    uint256 claimedCount;
     ERC20 token;
-    uint airdropAmount;
-    string secret; 
+    uint256 airdropAmount;
+    string secret;
 
-    constructor(ERC20 _token, uint _airdropAmount) {
+    constructor(ERC20 _token, uint256 _airdropAmount) {
         airdropAmount = _airdropAmount;
-        token = _token; 
+        token = _token;
         claimedCount = 0;
     }
 
@@ -22,24 +21,31 @@ contract TokenAirDrop is Ownable {
         secret = _secret;
     }
 
-    function _compareStrings(string memory a, string memory b) internal pure returns (bool) {
-        return (keccak256(abi.encodePacked((a))) == keccak256(abi.encodePacked((b))));
+    function _compareStrings(string memory a, string memory b)
+        internal
+        pure
+        returns (bool)
+    {
+        return (keccak256(abi.encodePacked((a))) ==
+            keccak256(abi.encodePacked((b))));
     }
 
-    function claim(uint256 _key, string calldata _identifier) external {  
-        require(uint256(keccak256(abi.encodePacked(secret, _identifier))) == _key, "Key does not match ID!");
-        
+    function claim(uint256 _key, string calldata _identifier) external {
+        require(
+            uint256(keccak256(abi.encodePacked(secret, _identifier))) == _key,
+            "Key does not match ID!"
+        );
+
         bool hasClaimed = false;
-        for (uint i = 0; i < claimedCount; i++) 
-        {
+        for (uint256 i = 0; i < claimedCount; i++) {
             if (_compareStrings(claimed[i], _identifier)) hasClaimed = true;
         }
         require(!hasClaimed, "You already claimed the tokens once!");
-        
+
         token.transferFrom(owner(), msg.sender, airdropAmount);
     }
 
-    function remainingTokens() external view returns(uint) {
+    function remainingTokens() external view returns (uint256) {
         return token.allowance(owner(), address(this));
     }
 }
